Separate plan code from context
